класс подсчёта статистики вынесен в отдельную папку вместе с необходимыми типами
