+++ src/mixin/statsCalculatorMixin.ts
@@ -2,7 +2,7 @@ import { defineComponent } from 'vue'
 
 import store from '../store/index'
 
-import GameStatsCalculator from '@/GameStatsCalculator'
+import GameStatsCalculator from '@/GameStatsCalculator/GameStatsCalculator'
 
 export default defineComponent({
   data() {
