+++ src/types/types.ts
@@ -1,3 +1,5 @@
+import { OvertimeStats } from '@/GameStatsCalculator/types'
+
 export type Operators = ['+', '-', '*', '/', '**']
 
 export type Operator = Operators[number]
@@ -53,27 +55,6 @@ export type GameStat = {
 export type GameStats = GameStat[]
 
 export type GameSessionStatus = 'ongoing' | 'ended';
-
-export type GameSessionRate = {
-  total: number
-  correct: number
-  incorrect: number
-}
-
-export type GameSession = {
-  id: string
-  stats: GameStats
-  start: Date
-  end: Date | null
-  rate: GameSessionRate
-  status: GameSessionStatus
-}
-
-export type GameSessionData = Partial<GameSession>
-
-export type OvertimeStats = {
-  sessions: GameSession[]
-}
 
 export type State = {
   previousRoute: string
